refactor: move Head to hooks
