+++ packages/gatsby-theme-project-portal/src/components/index.tsx
@@ -12,7 +12,6 @@ export type { CardProps } from "./Card"
 export { Cards } from "./Cards"
 export { NavbarLayout, Navbar } from "./Navbar"
 export { BottomBannerLayout, BottomBanner } from "./BottomBanner"
-export { SiteMetadata } from "./SiteMetadata"
 export { FooterLayout, Footer } from "./Footer"
 export { Contact } from "./Contact"
 export { ProjectTeam } from "./ProjectTeam"
