+++ packages/gatsby-theme-project-portal/src/hooks/index.tsx
@@ -1,3 +1,4 @@
 export { useProjectPortalConfig } from "./useProjectPortalConfig"
 export { useSiteMetadata } from "./useSiteMetadata"
 export { useStaticText } from "./useStaticText"
+export { Head } from "./Head"
