+++ packages/gatsby-theme-project-portal/src/templates/about-page.js
@@ -35,4 +35,4 @@ export const query = graphql`
   }
 `
 
-export { Head } from "../components/NewSiteMetadata"
+export { Head } from "../hooks"
