+++ packages/gatsby-theme-project-portal/src/templates/card-page.js
@@ -32,4 +32,4 @@ export const query = graphql`
   }
 `
 
-export { Head } from "../components/NewSiteMetadata"
+export { Head } from "../hooks"
