+++ packages/gatsby-theme-project-portal/src/templates/contact-page.js
@@ -26,4 +26,4 @@ export const query = graphql`
   }
 `
 
-export { Head } from "../components/NewSiteMetadata"
+export { Head } from "../hooks"
