+++ packages/gatsby-theme-project-portal/src/templates/project-detail-page.js
@@ -83,4 +83,4 @@ export const query = graphql`
   }
 `
 
-export { Head } from "../components/NewSiteMetadata"
+export { Head } from "../hooks"
