+++ packages/gatsby-theme-project-portal/src/templates/thank-you-page.js
@@ -23,4 +23,4 @@ export const query = graphql`
   }
 `
 
-export { Head } from "../components/NewSiteMetadata"
+export { Head } from "../hooks"
